Refacter HTTP request module
